refactor route tests
